Added a 'Sequential images' option to the Grid/Collection stitching plugin
Useful if you are working with images that you know you took in an overlapping o
rder (e.g. video from a microscope) but you don't have any kind of positional or
 grid information. Simply works by allowing the user to specify the range of ima
ges to be compared and then pairing images up,in the same way that the 'Unknown
Positions' option pairs all images and the grid option pairs up images expected
to overlap.