Stitching_: update status when fusion begins
Otherwise, in cases where Bio-Formats is used, the status bar
erroneously appears to still be populating OME metadata.

Ideally we would update the status in other areas too, especially when
registration is occurring. But unfortunately this would necessitate
adding showStatus calls in several places across copy-pasted code.

And it would be nice to show a progress bar for things like fusion,
since the fusion algorithm *does* have a general idea of how far along
it is. But this would require adding state to the Fusion class -- either
make it non-static, or pass more parameters to the static methods --
such that listeners can be registered and receive callbacks as the
fusion proceeds. Too involved for now...