Stitching: don't add multiple RoiPickers
When "Add tiles as ROIs" is selected, previously each time stitching
ran, RoiPicker would run as a plugin, causing "tool already exists"
messages to pop up on subsequent runs.

Added logic to check if the tool already exists. If it doesn't, it's run
as a plugin as before to enable and select it. If it does exist, it's
simply selected in the toolbar.