Fusion: display fusion in real time
When fusing blocks, an intermediate image will be displayed in real time
as the fused pixels are filled in. This should be helpful in determining
how much progress (if any) is actually being made during the fusion
process.

Only one time point/channel will be displayed at a time, and the
displayed image may not perfectly represent the final fused image - as
it will undergo rearrangement and post processing.

Note: as this does introduce operations per pixel during the fusion
process, this was benchmarked. On an 11 minute fusion, the time cost of
drawing the fusion in progress is < 20 seconds (under 3%).

Closes #8