POM: exclude erroneous gov.nih.imagej artifact
This version of ImageJ is obsolete and deprecated in favor of
net.imagej:ij. Furthermore, the fact that its artifactId is "imagej"
instead of "ij" causes both artifacts to appear on the classpath
simultaneously, which results in lots of duplicate classes at
incompatible versions. So it definitely needs to be excluded.