Fusion: correct resizing of regions
When splitting overlapping tiles, dimensions were only being resized
(e.g. to allow the tile with the most classes to "win") only when the
mins were unequal. This doesn't make sense - if the intervals were not
equal, they may need to be resized.

This should fix some black line issues.