Fusion: fix tile computation
There were several problems with the tile computation model. First, it
was a mistake to convert intervals to floats. They should be int-based,
and they should always be inclusive as this was the established logic.

Second, the presumption that position was not important when computing
overlapping intervals was faulty. It is important to fully considering
the ordering of points when deciding which intervals to create.

This commit updates the Interval model as indicated, and applies it
appropriately within tile computation.