Initial support for grid stitching from paths
Adds an extra grid order ('All files from path list')
to the 'Unknown position' and 'Sequential Images' grid types.

This option allows the tiles locations to be supplied via
a text file as a sequence of absolute paths in addition
to the current 'All files in directory' mode.
Unlike its directory-based counterpart, this option expects
all paths to be curated, hence disables the confirmation
dialog and performs no filtering based on file extensions.