GridType: make all choices available from a macro
There was already a pretty good effort to make sure that all the
"Order" choices were available from a macro. But the code was a bit
fragile and prone to fall out of sync. This commit updates things to
programmatically generate the allChoices list based on _all_ the
choices, no matter what they are, without hardcoding any indices.